chore: slack package export 변경
