+++ servers/mumu/src/handler/github-webhook/pull_request.ts
@@ -2,7 +2,9 @@ import type { Request, Response } from "express";
 import { pullRequestSchema, type PullRequest } from "@makers-devops/github";
 import { createPullRequestThread } from "../../slack";
 import { getPullRequestThreadKey } from "../../slack/key";
-import { deleteSlackThreadData, findSlackThread, slackClient } from "@makers-devops/slack";
+import { slackClient } from "@makers-devops/slack/client";
+import { findSlackThread } from "@makers-devops/slack/thread";
+import { deleteSlackThreadData } from "@makers-devops/slack";
 import { PR_닫힘 } from "@makers-devops/slack-blocks";
 import { assignReviewersAndAssignee } from "../../github";
 import { selectReviewers } from "../../github/review";
